pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yml", "*.csv", "*.txt", "*.html", "*.gz", "*.joblib", "**/*.sql"]

[tool.coverage.run]
parallel = true
source = [
]
context = '${CONTEXT}'

[tool.pytest.ini_options]
addopts = [
  "--strict-config",
  "--strict-markers",
]
xfail_strict = true
junit_family = "xunit2"
filterwarnings = [
    "error",
    'ignore::pytest.PytestDeprecationWarning:pytest_cov',
]

[tool.poetry]
name = "BM"
version = "0.1.0"
description = "Burning Man Stuff"
authors = ["peter <pelbaor@gmail.com>"]

[project]
name = "BM"
description = "Burning Man Stuff"
readme = "README.md"
requires-python = ">=3.11.6"
dynamic = ["version"]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
pandas = "*"
ipython = "*"
streamlit = "*"
openpyxl = "*"
altair = "*"
watchdog = "*"
matplotlib = "*"
statsmodels = "*"
scikit-learn = "*"
plotly = "*"
pillow-heif = "*"
python-dotenv = "*"

[project.optional-dependencies]
dev = [
]

test = []

[tool.setuptools.dynamic]
version = {attr = "bm.__version__"}
